sample/decoupled projects/file_handling/file_handling/requirements.txt
djangorestframework==3.11.0
Django==3.0.4
pymongo==3.10.1
mongoengine==0.19.1
pytest
coverage
google-cloud
pandas
google-cloud-storage
pyyaml